refactor: serializer/deserializer 관련 디렉터리 생성 및 이동
